+++ elements/other/order.json
@@ -1,6 +1,7 @@
 [
   "password",
   "image",
+  "image_advanced",
   "screencast",
   "balloons",
   "inspiration"
